requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-environ==0.11.2
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
Faker==19.13.0
kombu==5.3.3
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
